// web/src/app/core/api/generated/models/run-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { RunStatus } from '../models/run-status';
export interface RunResponse {
  createdAt?: string;
  executionTimeMs?: (number | string);
  finishedAt?: (string | null);
  id?: string;
  memoryUsedKb?: (number | string);
  problemId?: (number | string);
  startedAt?: (string | null);
  status?: RunStatus;
  stderr?: (string | null);
  stdout?: (string | null);
}